+++ contracts/ERC223/ERC223Receiver.sol
@@ -1,0 +1,5 @@
+pragma solidity ^0.4.18;
+
+interface ERC223Receiver { 
+    function tokenFallback(address _from, uint _value, bytes _data) external;
+}
