--- contracts/ERC223/Receiver_Interface.sol
@@ -1,9 +1,0 @@
-pragma solidity ^0.4.9;
-
-/*
- * Contract that is working with ERC223 tokens
- */
-
-interface ERC223Receiver {
-    function tokenFallback(address _from, uint _value, bytes _data) external pure;
-}
