+++ contracts/Migrations.sol
@@ -1,23 +1,23 @@
 pragma solidity ^0.4.19;
 
 contract Migrations {
-  address public owner;
-  uint public last_completed_migration;
+    address public owner;
+    uint public last_completed_migration;
 
-  modifier restricted() {
-    if (msg.sender == owner) _;
-  }
+    modifier restricted() {
+        if (msg.sender == owner) _;
+    }
 
-  function Migrations() public {
-    owner = msg.sender;
-  }
+    function Migrations() public {
+        owner = msg.sender;
+    }
 
-  function setCompleted(uint completed) restricted public {
-    last_completed_migration = completed;
-  }
+    function setCompleted(uint completed) restricted public {
+        last_completed_migration = completed;
+    }
 
-  function upgrade(address new_address) restricted public {
-    Migrations upgraded = Migrations(new_address);
-    upgraded.setCompleted(last_completed_migration);
-  }
+    function upgrade(address new_address) restricted public {
+        Migrations upgraded = Migrations(new_address);
+        upgraded.setCompleted(last_completed_migration);
+    } 
 }
